[FEAT(PlayerTest)] 추상클래스 Player 테스트
