+++ src/test/kotlin/domain/UserTest.kt
@@ -38,24 +38,4 @@ class UserTest {
         // then
         assertThat(actual).isEqualTo(10)
     }
-
-    @Test
-    fun `카드 한장을 추가한다`() {
-        // given
-        val user = User.create(
-            "산군" to
-                listOf<Card>(
-                    Card(Card.Shape.CLUBS, Card.Value.FIVE),
-                    Card(Card.Shape.DIAMONDS, Card.Value.FIVE),
-                ),
-        )
-        val card = Card(Card.Shape.SPADES, Card.Value.FIVE)
-
-        // when
-        user.addCard(card)
-        val actual = user.cards
-
-        // then
-        assertThat(actual).hasSize(3)
-    }
 }
